Moved ValueInterpolator to package numerics and fixed integer division.
